pyproject.toml
[project]
name = "openshell-deep-agent"
version = "0.1.0"
description = "General-purpose deep agent using OpenShell as the on-prem sandbox provider"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==26.1.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.6",
    "click==8.3.2",
    "cmdstanpy==1.3.0",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "deepagents>=0.4.11",
    "distro==1.9.0",
    "dotenv>=0.9.9",
    "et-xmlfile==2.0.0",
    "fastapi==0.135.3",
    "fonttools==4.62.1",
    "frozenlist==1.8.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "holidays==0.93",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "importlib-resources==6.5.2",
    "jiter==0.13.0",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "kiwisolver==1.5.0",
    "langchain==1.2.13",
    "langchain-classic==1.0.3",
    "langchain-community==0.4.1",
    "langchain-core==1.2.21",
    "langchain-experimental==0.4.1",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langchain-openai==1.1.12",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.1.3",
    "langgraph-checkpoint==4.0.1",
    "langgraph-cli[inmem]>=0.4.18",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.12",
    "langsmith==0.7.22",
    "marshmallow==3.26.2",
    "matplotlib==3.10.8",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "numpy==2.4.4",
    "openai==2.29.0",
    "opencv-python==4.13.0.92",
    "openpyxl==3.1.5",
    "openshell>=0.0.7",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==23.2",
    "pandas==3.0.1",
    "pdf2image==1.17.0",
    "pillow==12.1.1",
    "propcache==0.4.1",
    "prophet==1.3.0",
    "pulp==3.3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pyparsing==3.3.2",
    "pytesseract==0.3.13",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.24",
    "pyyaml==6.0.3",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.48",
    "stanio==0.5.1",
    "starlette==1.0.0",
    "tenacity==8.5.0",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uv==0.10.12",
    "uvicorn==0.44.0",
    "xgboost==3.2.0",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zstandard==0.25.0",
]

[tool.setuptools]
packages = ["src"]
